Fix handling of short packets
When transmitting short packets (shorter than FIFOThreshold), the
transmitter does not start automatically. Ensure the FIFO is flushed
after adding the data. This also causes DIO0 to be HIGH for a short
amount of time (and it could be missed), so DIO0 is handled with
interrupts also in TX mode.