Implemented setIdleMode
setIdleMode is used to turn the transmitter ON when no data needs to be
transmitted.
Code clean-up: several functions made private to avoid giving too many
possibilities to users.

TODO: implement an interrupt monitoring DIO5 to switch mode. It has
been hacked with a delay of 500ms but it should be fixed